[MASTERDIET-35] Post review fixes.
